posix specifics
