+++ src/config.c
@@ -36,19 +36,11 @@
 
 #include <tlspool/internal.h>
 
-#ifdef WINDOWS_PORT
-#include <windows.h>
-#endif
-
 //NOTYET// #include <libmemcached/memcached.h>
 
 //NOTYET// static LDAP *ldap_handle;
 
 //NOTYET// static struct memcached_st *cache;
-
-#ifdef WINDOWS_PORT
-char szPipename[1024];
-#endif
 
 static int kill_old_pid = 0;
 
@@ -101,25 +93,23 @@ enum VARS {
 };
 
 void cfg_setvar (char *item, int itemno, char *value);
-void cfg_pidfile (char *item, int itemno, char *value);
 void cfg_socketname (char *item, int itemno, char *value);
-void cfg_user (char *item, int itemno, char *value);
-void cfg_group (char *item, int itemno, char *value);
-void cfg_chroot (char *item, int itemno, char *value);
 void cfg_p11path (char *item, int itemno, char *value);
 void cfg_p11token (char *item, int itemno, char *value);
 void cfg_ldap (char *item, int itemno, char *value);
 void cfg_cachehost (char *item, int itemno, char *value);
+#ifndef WINDOWS_PORT
+void cfg_pidfile (char *item, int itemno, char *value);
+void cfg_user (char *item, int itemno, char *value);
+void cfg_group (char *item, int itemno, char *value);
+void cfg_chroot (char *item, int itemno, char *value);
+#endif
 
 struct cfgopt config_options [] = {
-	"daemon_pidfile",	cfg_pidfile,	CFGVAR_DAEMON_PIDFILE,
 	"socket_user",		cfg_setvar,	CFGVAR_SOCKET_USER,
 	"socket_group",		cfg_setvar,	CFGVAR_SOCKET_GROUP,
 	"socket_mode",		cfg_setvar,	CFGVAR_SOCKET_MODE,
 	"socket_name",		cfg_socketname,	CFGVAR_NONE,
-	"daemon_user",		cfg_user,	CFGVAR_NONE,
-	"daemon_group",		cfg_group,	CFGVAR_NONE,
-	"daemon_chroot",	cfg_chroot,	CFGVAR_NONE,
 	"pkcs11_path",		cfg_setvar,	CFGVAR_PKCS11_PATH,
 	"pkcs11_pin",		cfg_p11path,	CFGVAR_PKCS11_PIN,
 	"pkcs11_token",		cfg_p11token,	CFGVAR_NONE,
@@ -152,7 +142,12 @@ struct cfgopt config_options [] = {
 	"quantum_proof_authentication",cfg_setvar, CFGVAR_POSTQUANTUM_AUTH,
 	"quantum_proof_encryption",cfg_setvar,	CFGVAR_POSTQUANTUM_ENCRYPT,
 	"quantum_proof_handshake",cfg_setvar,	CFGVAR_POSTQUANTUM_HANDSHAKE,
-	//
+#ifndef WINDOWS_PORT
+	"daemon_pidfile",	cfg_pidfile,	CFGVAR_DAEMON_PIDFILE,
+	"daemon_user",		cfg_user,	CFGVAR_NONE,
+	"daemon_group",		cfg_group,	CFGVAR_NONE,
+	"daemon_chroot",	cfg_chroot,	CFGVAR_NONE,
+#endif /* WINDOWS_PORT */
 	NULL,			NULL,		CFGVAR_NONE
 };
 
@@ -212,7 +207,13 @@ struct var2val v2v_facility_flag [] = {
 	{ NULL, 0 }
 };
 
+static char *configvars [CFGVAR_LENGTH];
 
+#ifdef WINDOWS_PORT
+#include "config_windows.c"
+#else
+#include "config_posix.c"
+#endif
 
 static unsigned int parse_var2val (char *word, int wlen, struct var2val *patterns, unsigned int defaultvalue) {
 	if (word == NULL) {
@@ -331,9 +332,6 @@ void parse_cfgfile (char *filename, int kill_competition) {
 	fclose (cf);
 }
 
-
-static char *configvars [CFGVAR_LENGTH];
-
 void cfg_setvar (char *item, int itemno, char *value) {
 	if (configvars [itemno]) {
 		free (configvars [itemno]);
@@ -345,191 +343,6 @@ void cfg_setvar (char *item, int itemno, char *value) {
 	}
 #ifdef DEBUG
 	fprintf (stdout, "DEBUG: SETUP   %s AS %s\n", item, value);
-#endif
-}
-
-void unlink_pidfile (void) {
-#ifndef WINDOWS_PORT
-#ifndef CONFIG_PARSE_ONLY
-	unlink (configvars [CFGVAR_DAEMON_PIDFILE]);
-#endif
-#endif
-}
-
-void cfg_pidfile (char *item, int itemno, char *value) {
-#ifndef WINDOWS_PORT
-	static int fh = 0;
-	if (fh) {
-		fprintf (stderr, "You can specify only one PID file\n");
-		exit (1);
-	}
-	cfg_setvar (item, CFGVAR_DAEMON_PIDFILE, value);
-	atexit (unlink_pidfile);
-#ifndef CONFIG_PARSE_ONLY
-	fh = open (value, O_RDWR | O_CREAT, 0664);
-	char pidbuf [10];
-	if (fh < 0) {
-		perror ("Failed to open PID file");
-		exit (1);
-	}
-retry:
-	if (flock (fh, LOCK_EX | LOCK_NB) != 0) {
-		if (errno == EWOULDBLOCK) {
-			pid_t oldpid;
-			memset (pidbuf, 0, sizeof (pidbuf));
-			read (fh, pidbuf, sizeof (pidbuf)-1);
-			oldpid = atoi (pidbuf);
-			if (kill_old_pid) {
-				if (kill (oldpid, SIGHUP) == 0) {
-					fprintf (stderr, "Sent hangup to old daemon with PID %s\n", pidbuf);
-					sleep (1);
-				}
-				lseek (fh, 0, SEEK_SET);
-				errno = 0;
-				goto retry;
-			}
-			if (kill (oldpid, 0) != -1) {
-				fprintf (stderr, "Another daemon owns the PID file: process %s", pidbuf);
-				exit (1);
-			}
-		} else {
-			perror ("Failed to own the PID file");
-			exit (1);
-		}
-	}
-	snprintf (pidbuf, sizeof (pidbuf)-1, "%d\n", getpid ());
-	if (write (fh, pidbuf, strlen (pidbuf)) != strlen (pidbuf)) {
-		perror ("Failed to write all bytes to PID file");
-		exit (1);
-	}
-	ftruncate (fh, strlen (pidbuf));
-	fsync (fh);
-	//
-	// Note: The file remains open -- to sustain the flock on it
-	//
-#endif /* CONFIG_PARSE_ONLY */
-#endif /* WINDOWS_PORT */
-}
-
-void cfg_socketname (char *item, int itemno, char *value) {
-#ifdef WINDOWS_PORT
-	if (strlen (value) + 1 > sizeof (szPipename)) {
-		fprintf (stderr, "Socket path too long: %s\n", value);
-		exit (1);
-	}
-	strcpy (szPipename, value);
-#else
-	struct sockaddr_un sun;
-	int sox;
-#ifndef CONFIG_PARSE_ONLY
-#ifndef HAVE_SYSTEMD
-	uid_t me = getuid ();
-	gid_t my = getgid ();
-	if (strlen (value) + 1 > sizeof (sun.sun_path)) {
-		fprintf (stderr, "Socket path too long: %s\n", value);
-		exit (1);
-	}
-	strcpy (sun.sun_path, value);
-	sun.sun_family = AF_UNIX;
-	if (configvars [CFGVAR_DAEMON_PIDFILE]) {
-		//
-		// Note: Only be so kind to unlink when PID file is owned
-		//
-		unlink (value);
-	}
-	sox = socket (AF_UNIX, SOCK_STREAM, 0);
-	if (sox < 0) {
-		perror ("Failed to open UNIX socket");
-		exit (1);
-	}
-	if (bind (sox, (struct sockaddr *) &sun, SUN_LEN (&sun)) == -1) {
-		perror ("Failed to bind to UNIX socket");
-		exit (1);
-	}
-	if (listen (sox, 5) == -1) {
-		perror ("Failed to listen to UNIX socket");
-		exit (1);
-	}
-	//
-	// Now continue to set uid, gid, mode for the socket
-	//
-	if (configvars [CFGVAR_SOCKET_USER]) {
-		struct passwd *pwd = getpwnam (configvars [CFGVAR_SOCKET_USER]);
-		if (!pwd) {
-			fprintf (stderr, "Failed to find socket user %s\n", value);
-			exit (1);
-		}
-		me = pwd->pw_uid;
-	}
-	if (configvars [CFGVAR_SOCKET_GROUP]) {
-		struct group *grp = getgrnam (configvars [CFGVAR_SOCKET_GROUP]);
-		if (!grp) {
-			fprintf (stderr, "Failed to find socket group %s\n", value);
-			exit (1);
-		}
-		my = grp->gr_gid;
-	}
-	if (chown (value, me, my) != 0) {
-		perror ("Failed to change socket user/group");
-		exit (1);
-	}
-	if (configvars [CFGVAR_SOCKET_MODE]) {
-		int mode = strtoul (configvars [CFGVAR_SOCKET_MODE], NULL, 0);
-		if (chmod (value, mode) != 0) {
-			perror ("Failed to change socket mode");
-			exit (1);
-		}
-	}
-#else  /* HAVE_SYSTEMD */
-	if (sd_listen_fds (0) != 1) {
-		fprintf (stderr, "TLS Pool should have received one socket\n");
-		exit (1);
-	}
-	sox = SD_LISTEN_FDS_START + 0;
-#endif /* HAVE_SYSTEMD */
-	register_server_socket (sox);
-#endif /* CONFIG_PARSE_ONLY */
-#endif /* WINDOWS_PORT */
-}
-
-void cfg_user (char *item, int itemno, char *value) {
-#ifndef WINDOWS_PORT
-#ifdef DEBUG
-	fprintf (stdout, "DEBUG: DECLARE %s AS %s\n", item, value);
-#endif /* DEBUG */
-#ifndef CONFIG_PARSE_ONLY
-	struct passwd *pwd = getpwnam (value);
-	if (!pwd) {
-		fprintf (stderr, "Failed to find username %s\n", value);
-		exit (1);
-	}
-	setuid (pwd->pw_uid);
-#endif /* CONFIG_PARSE_ONLY */
-#endif /* WINDOWS_PORT */
-}
-
-void cfg_group (char *item, int itemno, char *value) {
-#ifndef WINDOWS_PORT
-#ifdef DEBUG
-	fprintf (stdout, "DEBUG: DECLARE %s AS %s\n", item, value);
-#endif /* DEBUG */
-#ifndef CONFIG_PARSE_ONLY
-	struct group *grp = getgrnam (value);
-	if (!grp) {
-		fprintf (stderr, "Failed to find group name %s\n", value);
-		exit (1);
-	}
-	setgid (grp->gr_gid);
-#endif /* CONFIG_PARSE_ONLY */
-#endif /* WINDOWS_PORT */
-}
-
-void cfg_chroot (char *item, int itemno, char *value) {
-#ifndef WINDOWS_PORT
-	if (chroot (value) != 0) {
-		perror ("Failed to chroot");
-		exit (1);
-	}
 #endif
 }
