+++ src/service.c
@@ -38,16 +38,9 @@
 #define WEOF ((wint_t)(0xFFFF))
 #endif
 
-#define PIPE_TIMEOUT 5000
-#define BUFSIZE 4096
-
 #define _tprintf printf
 #define _tmain main
 #endif /* WINDOWS_PORT */
-
-#ifdef WINDOWS_PORT
-extern char szPipename[1024];
-#endif
 
 /* The data stored in this module consists of lists of sockets to listen to
  * for connection setup and command exchange, but not data communication.
@@ -79,9 +72,6 @@ extern char szPipename[1024];
  */
 
 
-static struct soxinfo soxinfo [1024];
-static struct pollfd soxpoll [1024];
-static int num_sox = 0;
 static int stop_service = 0;
 static uint32_t facilities;
 
@@ -89,6 +79,8 @@ static struct callback cblist [1024];
 static struct callback *cbfree;
 static pthread_mutex_t cbfree_mutex = PTHREAD_MUTEX_INITIALIZER;
 
+static int os_send_command (struct command *cmd, int passfd);
+static void os_run_service ();
 
 /* Setup the service module.
  */
@@ -173,430 +165,15 @@ static void free_commands_by_clientfd (pool_handle_t clientfd) {
 	}
 }
 
-
-/* Register a socket.  It is assumed that first all server sockets register */
-void register_socket (pool_handle_t sox, uint32_t soxinfo_flags) {
-#ifndef WINDOWS_PORT
-	int flags = fcntl (sox, F_GETFD);
-	flags |= O_NONBLOCK;
-	fcntl (sox, F_SETFD, flags);
-	//TODO// if (soxinfo == NULL) {
-	//TODO// 	soxinfo = calloc ()
-	//TODO// }
-	if (num_sox == 1024) {
-		tlog (TLOG_UNIXSOCK, LOG_CRIT, "Cannot allocate more than 1024 server sockets");
-		exit (1);
-	}
-	soxpoll [num_sox].fd = sox;
-	soxpoll [num_sox].events = POLLIN;
-	soxpoll [num_sox].revents = 0;
-	soxinfo [num_sox].flags = soxinfo_flags;
-	soxinfo [num_sox].cbq = NULL;
-	num_sox++;
-#endif /* !WINDOWS_PORT */
-}
-
-
-void register_server_socket (pool_handle_t srvsox) {
-	register_socket (srvsox, SOF_SERVER);
-}
-
-
-void register_client_socket (pool_handle_t clisox) {
-	register_socket (clisox, SOF_CLIENT);
-}
-
-static void free_callbacks_by_clientfd (pool_handle_t clientfd);
-
-/* TODO: This may copy information back and thereby avoid processing in the
- * current loop passthrough.  No problem, poll() will show it once more. */
-static void unregister_client_socket_byindex (int soxidx) {
-#ifndef WINDOWS_PORT
-	pool_handle_t sox = soxpoll [soxidx].fd;
-	free_callbacks_by_clientfd (sox);
-	free_commands_by_clientfd (sox);
-	pinentry_forget_clientfd (sox);
-	lidentry_forget_clientfd (sox);
-	ctlkey_close_ctlfd (sox);
-	num_sox--;
-	if (soxidx < num_sox) {
-		memcpy (&soxinfo [soxidx], &soxinfo [num_sox], sizeof (*soxinfo));
-		memcpy (&soxpoll [soxidx], &soxpoll [num_sox], sizeof (*soxpoll));
-	}
-#endif /* !WINDOWS_PORT */
-}
-#ifdef WINDOWS_PORT
-#define CONNECTING_STATE 0
-#define READING_STATE 1
-#define INSTANCES 4
-#define PIPE_TIMEOUT 5000
-#define BUFSIZE 4096
-
-VOID DisconnectAndReconnect(DWORD);
-BOOL ConnectToNewClient(HANDLE, LPOVERLAPPED);
-void copy_tls_command(struct command *cmd, struct tlspool_command *tls_command);
-static void process_command (struct command *cmd);
-
-PIPEINST Pipe[INSTANCES];
-HANDLE hEvents[INSTANCES];
-
-#if defined(WINDOWS_PORT)
-static int socket_from_protocol_info (LPWSAPROTOCOL_INFOW lpProtocolInfo)
-{
-	return WSASocketW (FROM_PROTOCOL_INFO, FROM_PROTOCOL_INFO, FROM_PROTOCOL_INFO, lpProtocolInfo, 0, 0); 
-}
-#endif
-
-static int create_named_pipes (LPCTSTR lpszPipename)
-{
-   DWORD i, dwWait, cbRet, dwErr;
-   BOOL fSuccess;
-// The initial loop creates several instances of a named pipe
-// along with an event object for each instance.  An
-// overlapped ConnectNamedPipe operation is started for
-// each instance.
-
-   for (i = 0; i < INSTANCES; i++)
-   {
-
-   // Create an event object for this instance.
-
-      hEvents[i] = CreateEvent(
-         NULL,    // default security attribute
-         TRUE,    // manual-reset event
-         TRUE,    // initial state = signaled
-         NULL);   // unnamed event object
-
-      if (hEvents[i] == NULL)
-      {
-         printf("CreateEvent failed with %d.\n", GetLastError());
-         return 0;
-      }
-
-      Pipe[i].oOverlap.hEvent = hEvents[i];
-
-      Pipe[i].hPipeInst = CreateNamedPipe(
-         lpszPipename,            // pipe name
-         PIPE_ACCESS_DUPLEX |     // read/write access
-         FILE_FLAG_OVERLAPPED,    // overlapped mode
-         PIPE_TYPE_MESSAGE |      // message-type pipe
-         PIPE_READMODE_MESSAGE |  // message-read mode
-         PIPE_WAIT,               // blocking mode
-         INSTANCES,               // number of instances
-         BUFSIZE*sizeof(TCHAR),   // output buffer size
-         BUFSIZE*sizeof(TCHAR),   // input buffer size
-         PIPE_TIMEOUT,            // client time-out
-         NULL);                   // default security attributes
-
-      if (Pipe[i].hPipeInst == INVALID_HANDLE_VALUE)
-      {
-         printf("CreateNamedPipe failed with %d.\n", GetLastError());
-         return 0;
-      }
-
-   // Call the subroutine to connect to the new client
-
-      Pipe[i].fPendingIO = ConnectToNewClient(
-         Pipe[i].hPipeInst,
-         &Pipe[i].oOverlap);
-
-      Pipe[i].dwState = Pipe[i].fPendingIO ?
-         CONNECTING_STATE : // still connecting
-         READING_STATE;     // ready to read
-   }
-
-   while (1)
-   {
-   // Wait for the event object to be signaled, indicating
-   // completion of an overlapped read, write, or
-   // connect operation.
-
-      dwWait = WaitForMultipleObjects(
-         INSTANCES,    // number of event objects
-         hEvents,      // array of event objects
-         FALSE,        // does not wait for all
-         INFINITE);    // waits indefinitely
-
-   // dwWait shows which pipe completed the operation.
-
-      i = dwWait - WAIT_OBJECT_0;  // determines which pipe
-      if (i < 0 || i > (INSTANCES - 1))
-      {
-         printf("Index out of range.\n");
-         return 0;
-      }
-
-   // Get the result if the operation was pending.
-
-      if (Pipe[i].fPendingIO)
-      {
-         fSuccess = GetOverlappedResult(
-            Pipe[i].hPipeInst, // handle to pipe
-            &Pipe[i].oOverlap, // OVERLAPPED structure
-            &cbRet,            // bytes transferred
-            FALSE);            // do not wait
-
-         switch (Pipe[i].dwState)
-         {
-         // Pending connect operation
-            case CONNECTING_STATE:
-               if (! fSuccess)
-               {
-                   printf("Error %d.\n", GetLastError());
-                   return 0;
-               }
-               printf("Connected.\n");
-               Pipe[i].dwState = READING_STATE;
-               break;
-
-         // Pending read operation
-            case READING_STATE:
-               if (! fSuccess || cbRet == 0)
-               {
-                  printf("Error fSuccess = %d, cbRet = %d.\n", fSuccess, cbRet);
-                  DisconnectAndReconnect(i);
-                  continue;
-               }
-               printf("OK cbRet = %d.\n", cbRet);
-               Pipe[i].cbRead = cbRet;
-				struct command *cmd = allocate_command_for_clientfd (&Pipe[i]);
-				Pipe[i].chRequest.hPipe = Pipe[i].hPipeInst;
-				copy_tls_command (cmd, &Pipe[i].chRequest);
-				if (cmd->cmd.pio_ancil_type == ANCIL_TYPE_SOCKET) {
-					if (cmd->passfd == -1) {
-						//WRONG: no support for sockets
-						//HANDLE winsock = (HANDLE) WSASocket(FROM_PROTOCOL_INFO, FROM_PROTOCOL_INFO, FROM_PROTOCOL_INFO, &cmd->cmd.pio_ancil_data.pioa_socket, 0, 0);
-						//cmd->passfd = cygwin_attach_handle_to_fd(NULL, -1, winsock, NULL, GENERIC_READ | GENERIC_WRITE);
-						//tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Received file descriptor as %d, winsock = %d\n", cmd->passfd, winsock);
-						cmd->passfd = socket_from_protocol_info(&cmd->cmd.pio_ancil_data.pioa_socket);
-if (cmd->passfd == -1) printf("WSAGetLastError(void) = %d\n", WSAGetLastError());
-						tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Received file descriptor as %d\n", cmd->passfd);
-					} else {
-						//int superfd = (int) WSASocket(FROM_PROTOCOL_INFO, FROM_PROTOCOL_INFO, FROM_PROTOCOL_INFO, &cmd->cmd.pio_ancil_data.pioa_socket, 0, 0);
-						//tlog (TLOG_UNIXSOCK, LOG_ERR, "Received superfluous file descriptor as %d", superfd);
-						//close (superfd);
-					}
-				}
-				process_command (cmd);
-
-               break;
-
-            default:
-            {
-               printf("Invalid pipe state.\n");
-               return 0;
-            }
-         }
-      }
-
-   // The pipe state determines which operation to do next.
-
-      switch (Pipe[i].dwState)
-      {
-      // READING_STATE:
-      // The pipe instance is connected to the client
-      // and is ready to read a request from the client.
-
-         case READING_STATE:
-            fSuccess = ReadFile(
-               Pipe[i].hPipeInst,
-               &Pipe[i].chRequest,
-               sizeof (Pipe[i].chRequest),
-               &Pipe[i].cbRead,
-               &Pipe[i].oOverlap);
-
-         // The read operation completed successfully.
-
-            if (fSuccess && Pipe[i].cbRead != 0)
-            {
-				Pipe[i].fPendingIO = FALSE;
-
-               continue;
-            }
-
-         // The read operation is still pending.
-
-            dwErr = GetLastError();
-            if (! fSuccess && (dwErr == ERROR_IO_PENDING))
-            {
-               printf("read pending. %d\n", sizeof (Pipe[i].chRequest));
-               Pipe[i].fPendingIO = TRUE;
-               continue;
-            }
-            printf("The read failed with %d.\n", GetLastError());
-
-         // An error occurred; disconnect from the client.
-
-            DisconnectAndReconnect(i);
-            break;
-
-         default:
-         {
-            printf("Invalid pipe state.\n");
-            return 0;
-         }
-      }
-  }
-  return 0;
-}
-
-
-// DisconnectAndReconnect(DWORD)
-// This function is called when an error occurs or when the client
-// closes its handle to the pipe. Disconnect from this client, then
-// call ConnectNamedPipe to wait for another client to connect.
-
-VOID DisconnectAndReconnect(DWORD i)
-{
-// Disconnect the pipe instance.
-
-   if (! DisconnectNamedPipe(Pipe[i].hPipeInst) )
-   {
-      printf("DisconnectNamedPipe failed with %d.\n", GetLastError());
-   }
-
-// Call a subroutine to connect to the new client.
-
-   Pipe[i].fPendingIO = ConnectToNewClient(
-      Pipe[i].hPipeInst,
-      &Pipe[i].oOverlap);
-
-   Pipe[i].dwState = Pipe[i].fPendingIO ?
-      CONNECTING_STATE : // still connecting
-      READING_STATE;     // ready to read
-}
-
-// ConnectToNewClient(HANDLE, LPOVERLAPPED)
-// This function is called to start an overlapped connect operation.
-// It returns TRUE if an operation is pending or FALSE if the
-// connection has been completed.
-
-BOOL ConnectToNewClient(HANDLE hPipe, LPOVERLAPPED lpo)
-{
-   BOOL fConnected, fPendingIO = FALSE;
-
-// Start an overlapped connection for this pipe instance.
-   fConnected = ConnectNamedPipe(hPipe, lpo);
-
-// Overlapped ConnectNamedPipe should return zero.
-   if (fConnected)
-   {
-      printf("ConnectNamedPipe failed with %d.\n", GetLastError());
-      return 0;
-   }
-
-   switch (GetLastError())
-   {
-   // The overlapped connection in progress.
-      case ERROR_IO_PENDING:
-         fPendingIO = TRUE;
-         break;
-
-   // Client is already connected, so signal an event.
-
-      case ERROR_PIPE_CONNECTED:
-         if (SetEvent(lpo->hEvent))
-            break;
-
-   // If an error occurs during the connect operation...
-      default:
-      {
-         printf("ConnectNamedPipe failed with %d.\n", GetLastError());
-         return 0;
-      }
-   }
-   return fPendingIO;
-}
-
-static int np_send_command(struct tlspool_command *cmd) {
-	DWORD  cbToWrite, cbWritten;
-	OVERLAPPED overlapped;
-	BOOL fSuccess;
-
-	/* Send the request */
-	// Send a message to the pipe server.
-
-	cbToWrite = sizeof (struct tlspool_command);
-	_tprintf(TEXT("Sending %d byte cmd\n"), cbToWrite);
-
-	memset(&overlapped, 0, sizeof(overlapped));
-	overlapped.hEvent = CreateEvent(NULL, TRUE, FALSE, NULL);
-
-	fSuccess = WriteFile(
-		cmd->hPipe,                  // pipe handle
-		cmd,                    // cmd message
-		cbToWrite,              // cmd message length
-		NULL,                  // bytes written
-		&overlapped);            // overlapped
-
-	if (!fSuccess && GetLastError() == ERROR_IO_PENDING )
-	{
-printf ("DEBUG: Write I/O pending\n");
-		fSuccess = WaitForSingleObject(overlapped.hEvent, INFINITE) == WAIT_OBJECT_0;
-	}
-
-	if (fSuccess) {
-		fSuccess = GetOverlappedResult(cmd->hPipe, &overlapped, &cbWritten, TRUE);
-	}
-
-	if (!fSuccess)
-	{
-		_tprintf(TEXT("WriteFile to pipe failed. GLE=%d\n"), GetLastError());
-		errno = EPIPE;
-		return -1;
-	} else {
-printf ("DEBUG: Wrote %ld bytes to pipe\n", cbWritten);
-	}
-printf("DEBUG: Message sent to server, receiving reply as follows:\n");
-	return 0;
-}
-#endif /* WINDOWS_PORT */
-
 int send_command (struct command *cmd, int passfd) {
-#ifdef WINDOWS_PORT
-	cmd->cmd.pio_ancil_type = ANCIL_TYPE_NONE;
-	memset (&cmd->cmd.pio_ancil_data,
-			0,
-			sizeof (cmd->cmd.pio_ancil_data));
-	return !np_send_command(&cmd->cmd) ? 1 : 0;
-#else /* WINDOWS_PORT */
-	char anc [CMSG_SPACE(sizeof (int))];
-	struct iovec iov;
-	struct msghdr mh;
-	struct cmsghdr *cmsg;
+	int rc;
 
 	if (cmd == NULL) {
 		return 1;	// Success guaranteed when nobody is listening
 	}
 	assert (passfd == -1);	// Working passfd code retained but not used
-	memset (anc, 0, sizeof (anc));
-	memset (&iov, 0, sizeof (iov));
-	memset (&mh, 0, sizeof (mh));
-	iov.iov_base = &cmd->cmd;
-	iov.iov_len = sizeof (cmd->cmd);
-	mh.msg_iov = &iov;
-	mh.msg_iovlen = 1;
-	if (passfd >= 0) {
-		mh.msg_control = anc;
-		mh.msg_controllen = sizeof (anc);
-		cmsg = CMSG_FIRSTHDR (&mh);
-		cmsg->cmsg_level = SOL_SOCKET;
-		cmsg->cmsg_type = SCM_RIGHTS;
-		cmsg->cmsg_len = CMSG_LEN (sizeof (int));
-		* (int *) CMSG_DATA (cmsg) = passfd;
-	}
-	tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Sending command 0x%08x and fd %d to socket %d", cmd->cmd.pio_cmd, passfd, (int) cmd->clientfd);
-	if (sendmsg (cmd->clientfd, &mh, MSG_NOSIGNAL) == -1) {
-		//TODO// Differentiate behaviour based on errno?
-		perror ("Failed to send command");
-		cmd->claimed = 0;
-		return 0;
-	} else {
-		tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Sent command code 0x%08x", cmd->cmd.pio_cmd);
-		cmd->claimed = 0;
-		return 1;
-	}
-#endif /* WINDOWS_PORT */
+	cmd->claimed = 0;
+	return os_send_command(cmd, passfd);
 }
 
 /* Report success to the user.  Note that this function does not terminate
@@ -646,51 +223,6 @@ void send_error (struct command *cmd, int tlserrno, char *msg) {
 		perror ("Failed to send error reply");
 	}
 }
-
-
-#ifndef WINDOWS_PORT
-/* Receive a command.  Return nonzero on success, zero on failure. */
-int receive_command (pool_handle_t sox, struct command *cmd) {
-	int newfds [2];
-	int newfdcnt = 0;
-	char anc [CMSG_SPACE (sizeof (int))];
-	struct iovec iov;
-	struct msghdr mh = { 0 };
-	struct cmsghdr *cmsg;
-
-	iov.iov_base = &cmd->cmd;
-	iov.iov_len = sizeof (cmd->cmd);
-	mh.msg_iov = &iov;
-	mh.msg_iovlen = 1;
-	mh.msg_control = anc;
-	mh.msg_controllen = sizeof (anc);
-
-	if (recvmsg (sox, &mh, MSG_NOSIGNAL) == -1) {
-		//TODO// Differentiate behaviour based on errno?
-		perror ("Failed to receive command");
-		return 0;
-	}
-	tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Received command request code 0x%08x with cbid=%d over fd=%d", cmd->cmd.pio_cmd, cmd->cmd.pio_cbid, sox);
-
-	cmsg = CMSG_FIRSTHDR (&mh);
-	//TODO// It is more general to look at all FDs passed, close all 2+
-	if (cmsg && (cmsg->cmsg_len == CMSG_LEN (sizeof (int)))) {
-		if ((cmsg->cmsg_level == SOL_SOCKET) && (cmsg->cmsg_type == SCM_RIGHTS)) {
-			if (cmd->passfd == -1) {
-				cmd->passfd = * (int *) CMSG_DATA (cmsg);
-				tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Received file descriptor as %d", cmd->passfd);
-			} else {
-				int superfd = * (int *) CMSG_DATA (cmsg);
-				tlog (TLOG_UNIXSOCK, LOG_ERR, "Received superfluous file descriptor as %d", superfd);
-				close (superfd);
-			}
-		}
-		cmsg = CMSG_NXTHDR (&mh, cmsg);
-	}
-
-	return 1;
-}
-#endif /* !WINDOWS_PORT */
 
 void copy_tls_command(struct command *cmd, struct tlspool_command *tls_command) {
 	memcpy(&cmd->cmd, tls_command, sizeof(struct tlspool_command));
@@ -945,42 +477,6 @@ printf ("DEBUG: Processing callback command sent over fd=%d\n", cmd->clientfd);
 	}
 }
 
-
-/* Pickup on activity and process it.  Processing may mean a number of things:
- *  - to try an accept() on a server socket (ignoring it upon EAGAIN)
- *  - to trigger a thread that is hoping writing after EAGAIN
- *  - to read a message and further process it
- */
-#ifndef WINDOWS_PORT
-void process_activity (pool_handle_t sox, int soxidx, struct soxinfo *soxi, short int revents) {
-	if (revents & POLLOUT) {
-		//TODO// signal waiting thread that it may continue
-		tlog (TLOG_UNIXSOCK, LOG_CRIT, "Eekk!!  Could send a packet?!?  Unregistering client");
-		unregister_client_socket_byindex (soxidx);
-		tlspool_close_poolhandle (sox);
-	}
-	if (revents & POLLIN) {
-		if (soxi->flags & SOF_SERVER) {
-			struct sockaddr sa;
-			socklen_t salen = sizeof (sa);
-			pool_handle_t newsox = accept (sox, &sa, &salen);
-			if (newsox != INVALID_POOL_HANDLE) {
-				tlog (TLOG_UNIXSOCK, LOG_NOTICE, "Received incoming connection.  Registering it");
-				register_client_socket (newsox);
-			}
-		}
-		if (soxi->flags & SOF_CLIENT) {
-			struct command *cmd = allocate_command_for_clientfd (sox);
-			if (receive_command (sox, cmd)) {
-				process_command (cmd);
-			} else {
-				tlog (TLOG_UNIXSOCK, LOG_ERR, "Failed to receive command request");
-			}
-		}
-	}
-}
-#endif
-
 /* Request orderly hangup of the service.
  */
 void hangup_service (void) {
@@ -1005,32 +501,12 @@ void run_service (void) {
 		cblist [i].followup = NULL;
 		cbfree = &cblist [i];
 	}
-#ifdef WINDOWS_PORT
-	create_named_pipes ((LPCTSTR) szPipename);
-#else /* WINDOWS_PORT */
-	int polled;
-	tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Polling %d sockets numbered %d, %d, %d, ...", num_sox, soxpoll [0].fd, soxpoll [1].fd, soxpoll [2].fd);
-	while (polled = poll (soxpoll, num_sox, -1), polled > 0) {
-		tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Polled %d sockets, returned %d", num_sox, polled);
-		for (i=0; i<num_sox; i++) {
-			if (soxpoll [i].revents & (POLLHUP|POLLERR|POLLNVAL)) {
-				pool_handle_t sox = soxpoll [i].fd;
-				tlog (TLOG_UNIXSOCK, LOG_NOTICE, "Unregistering socket %d", sox);
-				unregister_client_socket_byindex (i);
-				close (sox);
-				continue;
-			} else if (soxpoll [i].revents) {
-				tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Socket %d has revents=%d", soxpoll [i].fd, soxpoll [i].revents);
-				process_activity (soxpoll [i].fd, i, &soxinfo [i], soxpoll [i].revents);
-			}
-		}
-		tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Polling %d sockets numbered %d, %d, %d, ...", num_sox, soxpoll [0].fd, soxpoll [1].fd, soxpoll [2].fd);
-	}
-	if (stop_service) {
-		tlog (TLOG_UNIXSOCK, LOG_NOTICE, "Service hangup in response to request");
-	} else {
-		tlog (TLOG_UNIXSOCK, LOG_DEBUG, "Polled %d sockets, returned %d", num_sox, polled);
-		perror ("Failed to poll for activity");
-	}
-#endif /* WINDOWS_PORT */
+	os_run_service ();
 }
+
+#ifdef WINDOWS_PORT
+#include "service_windows.c"
+#else
+#include "service_posix.c"
+#endif
+
